Move GlobalIdUtil to metadata normalizer
